final-project/main.py
import torch
import torch.nn as nn
import torch.optim as optim
import torch.nn.functional as F
from torch.utils.data import Dataset, DataLoader, random_split
import torchvision
from torchvision import transforms, datasets
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score, confusion_matrix, roc_curve, auc
from PIL import Image
import os
import time
from tqdm import tqdm
from tensorboard import SummaryWriter
import math


class ImagePreprocessor:
    """Handle image transformations and normalization"""
    
    def __init__(self, img_size=224, augment=True):
        self.img_size = img_size
        self.augment = augment
    
    def get_train_transforms(self):
        """Returns training transforms with augmentation"""
        pass
    
    def get_val_transforms(self):
        """Returns validation/test transforms without augmentation"""
        pass


class DatasetSplitter:
    """Manage train/validation/test splits"""
    
    def __init__(self, dataset_path, val_split=0.2):
        self.dataset_path = dataset_path
        self.val_split = val_split
    
    def load_datasets(self):
        """Load and split datasets"""
        pass
    
    def get_class_weights(self):
        """Calculate class weights for imbalanced dataset"""
        pass


class PneumoniaDataLoader:
    """Handle dataset loading, preprocessing, and batching"""
    
    def __init__(self, dataset_path, batch_size=32, img_size=224, val_split=0.2):
        self.dataset_path = dataset_path
        self.batch_size = batch_size
        self.img_size = img_size
        self.val_split = val_split
        self.preprocessor = ImagePreprocessor(img_size)
        self.splitter = DatasetSplitter(dataset_path, val_split)
    
    def create_data_loaders(self):
        """Create train, validation, and test data loaders"""
        pass
    
    def get_dataset_stats(self):
        """Get dataset statistics"""
        pass


class PatchEmbedding(nn.Module):
    """Convert images to patch embeddings for ViT"""
    
    def __init__(self, img_size=224, patch_size=16, in_channels=3, embed_dim=768):
        super().__init__()
        self.img_size = img_size
        self.patch_size = patch_size
        self.num_patches = (img_size // patch_size) ** 2
        self.embed_dim = embed_dim
        
        self.projection = nn.Conv2d(in_channels, embed_dim, kernel_size=patch_size, stride=patch_size)
    
    def forward(self, x):
        pass


class MultiHeadAttention(nn.Module):
    """Multi-head self-attention mechanism"""
    
    def __init__(self, embed_dim, num_heads, dropout=0.1):
        super().__init__()
        self.embed_dim = embed_dim
        self.num_heads = num_heads
        self.head_dim = embed_dim // num_heads
    
    def forward(self, x):
        pass


class MLP(nn.Module):
    """Feed-forward network for transformer"""
    
    def __init__(self, embed_dim, hidden_dim, dropout=0.1):
        super().__init__()
        self.fc1 = nn.Linear(embed_dim, hidden_dim)
        self.fc2 = nn.Linear(hidden_dim, embed_dim)
        self.dropout = nn.Dropout(dropout)
    
    def forward(self, x):
        pass


class TransformerBlock(nn.Module):
    """Transformer encoder block"""
    
    def __init__(self, embed_dim, num_heads, mlp_ratio=4, dropout=0.1):
        super().__init__()
        self.norm1 = nn.LayerNorm(embed_dim)
        self.attn = MultiHeadAttention(embed_dim, num_heads, dropout)
        self.norm2 = nn.LayerNorm(embed_dim)
        self.mlp = MLP(embed_dim, int(embed_dim * mlp_ratio), dropout)
    
    def forward(self, x):
        pass


class VisionTransformer(nn.Module):
    """Vision Transformer model implementation"""
    
    def __init__(self, img_size=224, patch_size=16, in_channels=3, num_classes=2, 
                 embed_dim=768, depth=12, num_heads=12, mlp_ratio=4, dropout=0.1):
        super().__init__()
        self.num_classes = num_classes
        self.embed_dim = embed_dim
        
        self.patch_embed = PatchEmbedding(img_size, patch_size, in_channels, embed_dim)
        num_patches = self.patch_embed.num_patches
        
        self.cls_token = nn.Parameter(torch.zeros(1, 1, embed_dim))
        self.pos_embed = nn.Parameter(torch.zeros(1, num_patches + 1, embed_dim))
        self.pos_drop = nn.Dropout(dropout)
        
        self.blocks = nn.ModuleList([
            TransformerBlock(embed_dim, num_heads, mlp_ratio, dropout)
            for _ in range(depth)
        ])
        
        self.norm = nn.LayerNorm(embed_dim)
        self.head = nn.Linear(embed_dim, num_classes)
    
    def forward(self, x):
        pass
    
    def get_num_params(self):
        """Return number of parameters"""
        pass


class CNNModel(nn.Module):
    """Convolutional Neural Network implementation"""
    
    def __init__(self, num_classes=2, dropout=0.5):
        super().__init__()
        self.num_classes = num_classes
        
        self.features = nn.Sequential(
            nn.Conv2d(3, 64, kernel_size=3, padding=1),
            nn.BatchNorm2d(64),
            nn.ReLU(inplace=True),
            nn.MaxPool2d(kernel_size=2, stride=2),
            
            nn.Conv2d(64, 128, kernel_size=3, padding=1),
            nn.BatchNorm2d(128),
            nn.ReLU(inplace=True),
            nn.MaxPool2d(kernel_size=2, stride=2),
            
            nn.Conv2d(128, 256, kernel_size=3, padding=1),
            nn.BatchNorm2d(256),
            nn.ReLU(inplace=True),
            nn.MaxPool2d(kernel_size=2, stride=2),
            
            nn.Conv2d(256, 512, kernel_size=3, padding=1),
            nn.BatchNorm2d(512),
            nn.ReLU(inplace=True),
            nn.MaxPool2d(kernel_size=2, stride=2),
        )
        
        self.classifier = nn.Sequential(
            nn.AdaptiveAvgPool2d((7, 7)),
            nn.Flatten(),
            nn.Linear(512 * 7 * 7, 4096),
            nn.ReLU(inplace=True),
            nn.Dropout(dropout),
            nn.Linear(4096, 1024),
            nn.ReLU(inplace=True),
            nn.Dropout(dropout),
            nn.Linear(1024, num_classes)
        )
    
    def forward(self, x):
        pass
    
    def get_num_params(self):
        """Return number of parameters"""
        pass


class MetricsCalculator:
    """Compute accuracy, precision, recall, F1-score"""
    
    def __init__(self):
        self.reset()
    
    def reset(self):
        """Reset metrics"""
        pass
    
    def update(self, predictions, targets):
        """Update metrics with new predictions"""
        pass
    
    def compute(self):
        """Compute final metrics"""
        pass
    
    def plot_confusion_matrix(self, predictions, targets, class_names):
        """Plot confusion matrix"""
        pass
    
    def plot_roc_curve(self, predictions, targets):
        """Plot ROC curve"""
        pass


class TrainingVisualizer:
    """Plot training curves and results"""
    
    def __init__(self, save_dir="./plots"):
        self.save_dir = save_dir
        os.makedirs(save_dir, exist_ok=True)
    
    def plot_training_curves(self, train_losses, val_losses, train_accs, val_accs):
        """Plot training and validation curves"""
        pass
    
    def plot_model_comparison(self, cnn_metrics, vit_metrics):
        """Compare CNN vs ViT performance"""
        pass
    
    def plot_training_time_comparison(self, cnn_time, vit_time):
        """Compare training times"""
        pass


class ModelTrainer:
    """Handle training loop, validation, and checkpointing"""
    
    def __init__(self, model, device, criterion, optimizer, scheduler=None, 
                 save_dir="./checkpoints"):
        self.model = model
        self.device = device
        self.criterion = criterion
        self.optimizer = optimizer
        self.scheduler = scheduler
        self.save_dir = save_dir
        self.metrics = MetricsCalculator()
        
        os.makedirs(save_dir, exist_ok=True)
    
    def train_epoch(self, train_loader):
        """Train for one epoch"""
        pass
    
    def validate_epoch(self, val_loader):
        """Validate for one epoch"""
        pass
    
    def train(self, train_loader, val_loader, num_epochs, early_stopping_patience=10):
        """Full training loop with early stopping"""
        pass
    
    def save_checkpoint(self, epoch, metrics, is_best=False):
        """Save model checkpoint"""
        pass
    
    def load_checkpoint(self, checkpoint_path):
        """Load model checkpoint"""
        pass


class ModelEvaluator:
    """Test set evaluation and performance analysis"""
    
    def __init__(self, model, device, criterion):
        self.model = model
        self.device = device
        self.criterion = criterion
        self.metrics = MetricsCalculator()
    
    def evaluate(self, test_loader):
        """Evaluate model on test set"""
        pass
    
    def detailed_analysis(self, test_loader, class_names):
        """Perform detailed analysis with visualizations"""
        pass
    
    def get_predictions(self, test_loader):
        """Get all predictions and targets"""
        pass


class ResultsLogger:
    """Save metrics and model artifacts"""
    
    def __init__(self, save_dir="./results"):
        self.save_dir = save_dir
        os.makedirs(save_dir, exist_ok=True)
    
    def log_training_results(self, model_name, training_history):
        """Log training history"""
        pass
    
    def log_evaluation_results(self, model_name, metrics):
        """Log evaluation metrics"""
        pass
    
    def save_model_summary(self, model_name, model, training_time, num_params):
        """Save model summary and statistics"""
        pass


class ModelComparator:
    """Compare CNN vs ViT performance"""
    
    def __init__(self, results_logger, visualizer):
        self.results_logger = results_logger
        self.visualizer = visualizer
    
    def compare_models(self, cnn_results, vit_results):
        """Compare two models comprehensively"""
        pass
    
    def generate_comparison_report(self, cnn_results, vit_results):
        """Generate detailed comparison report"""
        pass


class ExperimentRunner:
    """Orchestrate training of both models"""
    
    def __init__(self, config):
        self.config = config
        self.device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
        self.results_logger = ResultsLogger()
        self.visualizer = TrainingVisualizer()
        self.comparator = ModelComparator(self.results_logger, self.visualizer)
    
    def setup_data(self):
        """Setup data loaders"""
        pass
    
    def create_models(self):
        """Create CNN and ViT models"""
        pass
    
    def train_cnn(self):
        """Train CNN model"""
        pass
    
    def train_vit(self):
        """Train ViT model"""
        pass
    
    def evaluate_models(self):
        """Evaluate both models on test set"""
        pass
    
    def run_experiment(self):
        """Run complete experiment"""
        pass


def main():
    """Main execution function"""
    config = {
        'dataset_path': './chest_xray',
        'batch_size': 32,
        'img_size': 224,
        'num_epochs': 50,
        'learning_rate': 1e-4,
        'val_split': 0.2,
        'early_stopping_patience': 10,
        'save_dir': './results',
        
        # ViT specific
        'vit_patch_size': 16,
        'vit_embed_dim': 768,
        'vit_depth': 12,
        'vit_num_heads': 12,
        
        # CNN specific
        'cnn_dropout': 0.5,
    }
    
    print("Starting Pneumonia Classification Experiment...")
    print(f"Device: {torch.device('cuda' if torch.cuda.is_available() else 'cpu')}")
    
    experiment = ExperimentRunner(config)
    experiment.run_experiment()


if __name__ == "__main__":
    main()